Add translations for Flatpickr.js
Once Alchemy switches to using flatpickr as the default library for
date and time inputs, it will be very handy to have some translations.

Downloaded via https://cdnjs.com/libraries/flatpickr as there is no
canonical bundle.

Flatpickr does handle which day of the week the week starts in different
languages, but its localization does not know whether to show
24-hour-clocks. Wikipedia essentially says: 24 hour everywhere except
for the English-speaking world.