Do not change config.available_locales in install generator
This bit of code changes the host's app `application.rb` file,
restricting the available locales to ones that are not English. I have
an app that has both English and German as locales, and I have to delete
the added line every time I run the installer. I think it should stay
the host app's responsibility to choose what locales are good for it.